backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-jose[cryptography]
bcrypt==4.0.1
passlib==1.7.4
pydantic
sqlalchemy
pandas
psycopg2-binary
